+++ lang/src/main/kotlin/org/partiql/lang/compiler/PartiQLCompilerDefault.kt
@@ -14,7 +14,7 @@
 
 package org.partiql.lang.compiler
 
-import org.partiql.annotations.PartiQLExperimental
+import org.partiql.annotations.ExperimentalPartiQLCompilerPipeline
 import org.partiql.lang.domains.PartiqlAst
 import org.partiql.lang.domains.PartiqlLogical
 import org.partiql.lang.domains.PartiqlLogicalResolved
@@ -43,7 +43,7 @@ import org.partiql.lang.planner.EvaluatorOptions
 import org.partiql.lang.planner.PartiQLPlanner
 import org.partiql.lang.types.TypedOpParameter
 
-@PartiQLExperimental
+@ExperimentalPartiQLCompilerPipeline
 internal class PartiQLCompilerDefault(
     private val evaluatorOptions: EvaluatorOptions,
     private val customTypedOpParameters: Map<String, TypedOpParameter>,