+++ lang/src/main/kotlin/org/partiql/lang/eval/physical/PhysicalBexprToThunkConverter.kt
@@ -2,7 +2,7 @@ package org.partiql.lang.eval.physical
 
 import com.amazon.ionelement.api.BoolElement
 import com.amazon.ionelement.api.MetaContainer
-import org.partiql.annotations.PartiQLExperimental
+import org.partiql.annotations.ExperimentalWindowFunctions
 import org.partiql.lang.ast.SourceLocationMeta
 import org.partiql.lang.domains.PartiqlPhysical
 import org.partiql.lang.eval.ExprValue
@@ -294,8 +294,7 @@ internal class PhysicalBexprToThunkConverter(
         CompiledSortKey(comp, value)
     }
 
-    // TODO: Remove from experimental once https://github.com/partiql/partiql-docs/issues/31 is resolved and a RFC is approved
-    @PartiQLExperimental
+    @OptIn(ExperimentalWindowFunctions::class)
     override fun convertWindow(node: PartiqlPhysical.Bexpr.Window): RelationThunkEnv {
         val source = this.convert(node.source)