+++ lang/src/main/kotlin/org/partiql/lang/eval/physical/operators/WindowRelationalOperatorFactory.kt
@@ -1,12 +1,11 @@
 package org.partiql.lang.eval.physical.operators
 
-import org.partiql.annotations.PartiQLExperimental
+import org.partiql.annotations.ExperimentalWindowFunctions
 import org.partiql.lang.domains.PartiqlPhysical
 import org.partiql.lang.eval.physical.SetVariableFunc
 import org.partiql.lang.eval.physical.window.WindowFunction
 
-// TODO: Remove from experimental once https://github.com/partiql/partiql-docs/issues/31 is resolved and a RFC is approved
-@PartiQLExperimental
+@ExperimentalWindowFunctions
 abstract class WindowRelationalOperatorFactory(name: String) : RelationalOperatorFactory {
 
     final override val key: RelationalOperatorFactoryKey = RelationalOperatorFactoryKey(RelationalOperatorKind.WINDOW, name)
@@ -21,7 +20,7 @@ abstract class WindowRelationalOperatorFactory(name: String) : RelationalOperato
     ): RelationExpression
 }
 
-@PartiQLExperimental
+@ExperimentalWindowFunctions
 class CompiledWindowFunction(
     val func: WindowFunction,
     val parameters: List<ValueExpression>,